test: adding tests for UpdateNoteLinks (wip)
