+++ cmd/root.go
@@ -8,9 +8,10 @@ import (
 )
 
 var rootCmd = &cobra.Command{
-	Use:   "obs",
-	Short: "obs - CLI to open, search, move, create and update notes",
-	Long:  "obs - CLI to open, search, move, create and update notes",
+	Use:     "obs",
+	Short:   "obs - CLI to open, search, move, create and update notes",
+	Version: "v0.1.3",
+	Long:    "obs - CLI to open, search, move, create and update notes",
 }
 
 func Execute() {
